Refactor: 도메인 변경 (review -> animation)
