+++ src/main/java/com/otakumap/domain/animation/DTO/AnimationResponseDTO.java
@@ -1,0 +1,28 @@
+package com.otakumap.domain.animation.DTO;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+import java.util.List;
+
+public class AnimationResponseDTO {
+    @Builder
+    @Getter
+    @NoArgsConstructor
+    @AllArgsConstructor
+    public static class AnimationResultDTO {
+        Long animationId;
+        String name;
+    }
+
+    @Builder
+    @Getter
+    @NoArgsConstructor
+    @AllArgsConstructor
+    public static class AnimationResultListDTO {
+        List<AnimationResultDTO> animations;
+        Integer listSize;
+    }
+}
