+++ src/main/java/com/otakumap/domain/animation/service/AnimationQueryService.java
@@ -1,0 +1,9 @@
+package com.otakumap.domain.animation.service;
+
+import com.otakumap.domain.animation.entity.Animation;
+
+import java.util.List;
+
+public interface AnimationQueryService {
+    List<Animation> searchAnimation(String keyword);
+}
